Refactor to importer to own component and function
